Completed tests on readable/writable streams
