+++ tests/Adapter/Amp/EventLoopTest.php
@@ -12,9 +12,9 @@ class EventLoopTest extends \M6WebTest\Tornado\EventLoopTest
         return new Amp\EventLoop();
     }
 
-    public function testReadableStream($expectedSequence = '')
+    public function testStreamShouldReadFromWritable($expectedSequence = '')
     {
         // Because AMP doesn't trigger write callback immediately
-        parent::testReadableStream('W0R0W12345R12R34R5W6R6R');
+        parent::testStreamShouldReadFromWritable('W0R0W12345R12R34R5W6R6R');
     }
 }
