+++ tests/Adapter/Tornado/SynchronousEventLoopTest.php
@@ -30,9 +30,9 @@ class SynchronousEventLoopTest extends \M6WebTest\Tornado\EventLoopTest
         parent::testPromiseRaceShouldRejectIfFirstSettledPromiseRejects(1);
     }
 
-    public function testReadableStream($expectedSequence = '')
+    public function testStreamShouldReadFromWritable($expectedSequence = '')
     {
         // Never wait…
-        parent::testReadableStream('W0W12345W6R01R23R45R6R');
+        parent::testStreamShouldReadFromWritable('W0W12345W6R01R23R45R6R');
     }
 }
